More path modifications
