+++ code/game-server-v2/main.ts
@@ -4,23 +4,16 @@ import cors from "cors";
 import express from "express";
 import http from "http";
 import { Server, Socket } from "socket.io";
-import connectDb from "./db/db";
-import { GameServerRegisterModel, IGameServerRegisterEntry } from "./db/models";
+import connectDb from "./utils";
 import { log_attention, log_event, log_notice, log_warning } from "./utils";
 import * as fs from "fs";
 import * as path from "path";
 import { Player } from "./player";
-import { SideId } from "../beastly-brawl-showdown/imports/simulator/core/side";
-import { COMMON_MONSTER_POOL } from "../beastly-brawl-showdown/imports/simulator/data/common/common_monster_pool";
-import { log } from "console";
-import { EntryID } from "../beastly-brawl-showdown/imports/simulator/core/utils";
-import { TargetingMethod } from "../beastly-brawl-showdown/imports/simulator/core/action/targeting";
-import {
-  ChooseMove,
-  Roll,
-} from "../beastly-brawl-showdown/imports/simulator/core/notice/notice";
-import { match } from "assert";
+import { SideId } from "../simulator/core/side";
+import { COMMON_MONSTER_POOL } from "../simulator/data/common/common_monster_pool";
+import { TargetingMethod } from "../simulator/core/action/targeting";
 import { Match, MatchType } from "./match";
+import { GameServerRegistryModel, IGameServerRegistryEntry } from "../server-locator/src/models/game_server_register";
 
 type ServerConfig = {
   serverIp: string;
@@ -49,28 +42,24 @@ async function main(config: ServerConfig) {
   log_notice("Connect to database...");
   await connectDb();
   log_notice("Register to global records...");
-  const existingRecordCount = await GameServerRegisterModel.countDocuments({
+  const existingRecordCount = await GameServerRegistryModel.countDocuments({
     serverNumber: config.serverNumber,
   });
   if (existingRecordCount > 0) {
-    console.log(
-      `Exsting records found with server number <${config.serverNumber}>: ${existingRecordCount}`
-    );
+    console.log(`Exsting records found with server number <${config.serverNumber}>: ${existingRecordCount}`);
     if (!config.overrideExistingRecordOnStartup) {
       throw new Error("A record already exists, room could not be registered.");
     }
   }
-  const updatedRecord =
-    await GameServerRegisterModel.findOneAndUpdate<IGameServerRegisterEntry>(
-      { serverNumber: config.serverNumber },
-      {
-        serverNumber: config.serverNumber,
-        serverUrl:
-          config.serverIp.toString() + ":" + config.serverPort.toString(),
-        lastUpdated: new Date(),
-      },
-      { upsert: true, new: true }
-    );
+  const updatedRecord = await GameServerRegistryModel.findOneAndUpdate<IGameServerRegistryEntry>(
+    { serverNumber: config.serverNumber },
+    {
+      serverNumber: config.serverNumber,
+      serverUrl: config.serverIp.toString() + ":" + config.serverPort.toString(),
+      lastUpdated: new Date(),
+    },
+    { upsert: true, new: true }
+  );
   log_notice("New Record:\n" + JSON.stringify(updatedRecord));
   log_notice("Registered to records.");
 
@@ -118,9 +107,7 @@ async function main(config: ServerConfig) {
     // hostName: string;
   };
   hostChannel.use((socket, next) => {
-    log_event(
-      `Host attempted to join with ${JSON.stringify(socket.handshake.auth)}`
-    );
+    log_event(`Host attempted to join with ${JSON.stringify(socket.handshake.auth)}`);
     const auth = socket.handshake.auth as HostChannelAuth;
     /// for now always accept the host name
     // if (!auth.hostName) {
@@ -144,10 +131,7 @@ async function main(config: ServerConfig) {
       log_event("Room requested.");
       // TODO prevent multiple rooms at the same time
       try {
-        const { roomId: roomId, joinCode: joinCode } = gameServer.createRoom(
-          socket.id,
-          playerChannel
-        );
+        const { roomId: roomId, joinCode: joinCode } = gameServer.createRoom(socket.id, playerChannel);
 
         socket.emit("request-room_response", {
           roomId: roomId,
@@ -174,9 +158,7 @@ async function main(config: ServerConfig) {
         playerChannel.to(player.socketId).emit("game-started"); // Clients can now start monster selection
       });
 
-      log_notice(
-        `All players in room ${msg.roomId} have been notified to start the game.`
-      );
+      log_notice(`All players in room ${msg.roomId} have been notified to start the game.`);
     });
   });
 
@@ -218,18 +200,14 @@ async function main(config: ServerConfig) {
       res.send(checkResult);
       return;
     }
-    checkResult.isDisplayNameValid = !gameServer.rooms
-      .get(roomId)
-      ?.hasPlayer(req.body.displayName);
+    checkResult.isDisplayNameValid = !gameServer.rooms.get(roomId)?.hasPlayer(req.body.displayName);
 
     log_notice(`Player auth check result:\n${JSON.stringify(checkResult)}`);
     res.send(checkResult);
   });
 
   playerChannel.use((socket, next) => {
-    log_event(
-      `Player attempted to join with ${JSON.stringify(socket.handshake.auth)}`
-    );
+    log_event(`Player attempted to join with ${JSON.stringify(socket.handshake.auth)}`);
     const auth = socket.handshake.auth as PlayerChannelAuth;
 
     if (!auth.joinCode) {
@@ -267,26 +245,14 @@ async function main(config: ServerConfig) {
       next(new Error("Invalid credentials"));
     }
 
-    log_event(
-      `Join code <${auth.joinCode}> is valid. From <${auth.displayName}>. Socket id = ${socket.id}`
-    );
+    log_event(`Join code <${auth.joinCode}> is valid. From <${auth.displayName}>. Socket id = ${socket.id}`);
 
-    const playerNameList =
-      gameServer.rooms
-        .get(roomId)
-        ?.players.map((player) => player.displayName) ?? [];
+    const playerNameList = gameServer.rooms.get(roomId)?.players.map((player) => player.displayName) ?? [];
 
-    console.log(
-      "Update player list",
-      playerNameList,
-      "to",
-      gameServer.rooms.get(roomId)!.hostSocketId
-    );
+    console.log("Update player list", playerNameList, "to", gameServer.rooms.get(roomId)!.hostSocketId);
     console.log("Player socket: ", socket.data);
 
-    hostChannel
-      .to(gameServer.rooms.get(roomId)!.hostSocketId)
-      .emit("player-set-changed", playerNameList);
+    hostChannel.to(gameServer.rooms.get(roomId)!.hostSocketId).emit("player-set-changed", playerNameList);
     next();
   });
 
@@ -318,9 +284,7 @@ async function main(config: ServerConfig) {
       player.isReady = true;
 
       // Check if all players are ready
-      const allReady = Array.from(room.players.values()).every(
-        (p) => p.isReady
-      );
+      const allReady = Array.from(room.players.values()).every((p) => p.isReady);
       if (!allReady) {
         log_notice("Waiting for all players to submit their monsters...");
         return;
@@ -356,9 +320,7 @@ async function main(config: ServerConfig) {
       const player = socket.data.player as Player;
       const room = gameServer.rooms.get(player.roomId!);
       if (!room) return;
-      const match = room.tournamentManager.matches.find(
-        (m) => m.player1 === player || m.player2 === player
-      );
+      const match = room.tournamentManager.matches.find((m) => m.player1 === player || m.player2 === player);
       if (!match) return;
 
       match.submitRoll(player);
@@ -375,9 +337,7 @@ async function main(config: ServerConfig) {
       const room = gameServer.rooms.get(player.roomId!);
       if (!room) return;
 
-      const match = room.tournamentManager.matches.find(
-        (m) => m.player1 === player || m.player2 === player
-      );
+      const match = room.tournamentManager.matches.find((m) => m.player1 === player || m.player2 === player);
       if (!match) return;
       const [player1, player2] = [match.player1, match.player2];
 
@@ -386,21 +346,11 @@ async function main(config: ServerConfig) {
 
       switch (moveId) {
         case "defend":
-          match.submitMove(
-            player,
-            moveId,
-            targetMethod as TargetingMethod,
-            sourceSide as SideId
-          );
+          match.submitMove(player, moveId, targetMethod as TargetingMethod, sourceSide as SideId);
           break;
         case "attack-normal":
           const targetSide = sourceSide === 1 ? 0 : 1;
-          match.submitMove(
-            player,
-            moveId,
-            targetMethod as TargetingMethod,
-            targetSide as SideId
-          );
+          match.submitMove(player, moveId, targetMethod as TargetingMethod, targetSide as SideId);
           break;
       }
 
@@ -430,11 +380,7 @@ async function main(config: ServerConfig) {
   });
 
   httpServer.listen(config.serverPort, () => {
-    log_notice(
-      `Socket.IO server running on ${
-        config.serverIp.toString() + ":" + config.serverPort.toString()
-      }. <CTRL+C> to shutdown.`
-    );
+    log_notice(`Socket.IO server running on ${config.serverIp.toString() + ":" + config.serverPort.toString()}. <CTRL+C> to shutdown.`);
     //#endregion
 
     //#region IO
