+++ code/game-server-v2/player.ts
@@ -1,4 +1,4 @@
-import { MonsterTemplate } from "../beastly-brawl-showdown/imports/simulator/core/monster/monster_template";
+import { MonsterTemplate } from "../simulator/core/monster/monster_template";
 import { RoomId } from "./types";
 
 export class Player {
