+++ code/game-server-v2/utils.ts
@@ -1,5 +1,4 @@
-// TODO move to general utils
-
+//#region Logging
 /// See: https://stackoverflow.com/questions/4842424/list-of-ansi-color-escape-sequences
 export function log_notice(val: any) {
   console.log("\x1b[38;5;0m[Notice]\x1b[0m", val);
@@ -16,3 +15,19 @@ export function log_attention(val: any) {
 export function log_event(val: any) {
   console.log("\x1b[38;5;6m[Event]\x1b[0m", val);
 }
+//#endregion
+
+//#region DB
+import mongoose from "mongoose";
+const MONGO_URI = "mongodb://localhost:27017/test";
+
+export default async function connectDb() {
+  try {
+    mongoose.connect(MONGO_URI);
+    log_notice("Connected to MongoDB.");
+  } catch (error) {
+    console.error("MongoDB connection error:", error);
+    process.exit(1);
+  }
+}
+//#endregion
